IPUToolkit.jl

Index

IPUToolkit.IPUCompiler.DISABLE_PRINT

IPUToolkit.IPUCompiler.KEEP_LLVM_FILES

IPUToolkit.IPUCompiler.POPC_FLAGS

IPUToolkit.IPUCompiler.PROGRESS_SPINNER

Base.similar

IPUToolkit.IPUCompiler.add_vertex

IPUToolkit.Poplar.get_ipu_device

IPUToolkit.Poplar.get_ipu_devices

IPUToolkit.IPUCompiler.@ipucycles

IPUToolkit.IPUCompiler.@ipuelapsed

IPUToolkit.IPUCompiler.@ipuprint

IPUToolkit.IPUCompiler.@ipuprintf

IPUToolkit.IPUCompiler.@ipuprintln

IPUToolkit.IPUCompiler.@ipushow

IPUToolkit.IPUCompiler.@ipushowcycles

Types

Functions and macros

Base.similar — Method

similar(
    graph::Poplar.Graph,
    tensor::Poplar.TensorAllocated,
    [type::DataType],
    [debug::String]
) -> Poplar.TensorAllocated

Adds to graph a variable tensor with the same shape as tensor. If a type (this is a Julia type, like Float32 or Int32) argument is not passed, the same element type as tensor will be automatically used. An optional debug context can also be passed, as a String.

IPUToolkit.Poplar.get_ipu_device — Function

get_ipu_device(hint::Union{AbstractVector{<:Integer},Integer}=0)

Similar to get_ipu_devices, but request exactly one IPU device. If it can attach to a device, return that pointer only (not in a vector, like get_ipu_devices), otherwise return nothing. You can release the device with Poplar.DeviceDetach.

The optional argument hint suggests to which device IDs to try and attach. It can have different types:

if of type Integer, try to attach to n devices, starting from the one with index hint. The default is hint=0;

if of type AbstractVector, try to attach to n devices from that list of IDs.

IPUToolkit.Poplar.get_ipu_devices — Function

get_ipu_devices(n::Int, hint::Union{AbstractVector{<:Integer},Integer}=0)

Try to attach to n IPU devices, returns a vector of the pointers to the devices successfully attached to. You can release them with Poplar.DeviceDetach (note that this function takes a single pointer as input, so you have to use broadcasting Poplar.DeviceDetach.(devices) to release a vector of pointers).

The second optional argument hint suggests to which device IDs to try and attach. It can have different types:

if of type Integer, try to attach to n devices, starting from the one with index hint. The default is hint=0;

if of type AbstractVector, try to attach to n devices from that list of IDs.

See get_ipu_device for requesting exactly one IPU device.

IPUToolkit.IPUCompiler.add_vertex — Function

add_vertex(graph::Poplar.GraphAllocated,
           compute_set_or_program::Union{Poplar.ComputeSetAllocated, Poplar.ProgramSequenceAllocated},
           [tiles::Union{Integer,AbstractVector{<:Integer}},]
           codelet::Function,
           args::Poplar.TensorAllocated...) -> Nothing

Add the codelet function codelet created with @codelet to graph, using the tensors args as arguments. The function codelet must have exactly one method, no more, no less. The second argument can be either the program or the compute set to which to add the new vertex/vertices. If a program is passed, a new compute set will be automatically created.

add_vertex also evenly maps all tensors and vertices across all tiles, which can be either a single tile ID or an AbstractVector of IDs and defaults to single tile 0 if this argument is omitted. Note that all argument tensors args must be longer than or equal to the number of tiles. If you want to have better control over tile mapping, do not use add_vertex.

IPUToolkit.IPUCompiler.@ipucycles — Macro

@ipucycles ex
@ipucycles "description" ex

Print from inside a codelet the number of cycles spent to compute the expression ex. The corresponding time can be obtained by dividing the number of cycles by the clock frequency of the the tile, which you can get with Poplar.TargetGetTileClockFrequency(target) outside of the codelet. The optional argument description, a literal String, can be used to print also a label to identify the timed expression. A label is added automatically by @ipushowcycles.

See also @ipuelapsed.

This macro can be made no-op completely by setting

IPUToolkit.IPUCompiler.DISABLE_PRINT[] = true

IPUToolkit.IPUCompiler.@ipuelapsed — Macro

@ipuelapsed ex

Return number of cycles spent to compute the expression ex. The corresponding time can be obtained by dividing the number of cycles by the clock frequency of the the tile, which you can get with Poplar.TargetGetTileClockFrequency(target) outside of the codelet.

See also @ipucycles, @ipushowcycles.

IPUToolkit.IPUCompiler.@ipuprint — Macro

@ipuprint(xs...)
@ipuprintln(xs...)

Print a textual representation of values xs to standard output from the IPU. The functionality builds on @ipuprintf, and is intended as a more use friendly alternative of that API. However, that also means there's only limited support for argument types, handling 16/32/64 signed and unsigned integers, 32 and 64-bit floating point numbers, Cchars and pointers. For more complex output, use @ipuprintf directly.

Limited string interpolation is also possible:

    @ipuprint("Hello, World ", 42, "\n")
    @ipuprint "Hello, World $(42)\n"

Printing can be completely disabled by setting

IPUToolkit.IPUCompiler.DISABLE_PRINT[] = true

IPUToolkit.IPUCompiler.@ipuprintf — Macro

@ipuprintf("%Fmt", args...)

Print a formatted string in device context on the host standard output.

Note that this is not a fully C-compliant printf implementation.

Also beware that it is an untyped, and unforgiving printf implementation. Type widths need to match, eg. printing a 64-bit Julia integer requires the %ld formatting string.

Printing can be completely disabled by setting

IPUToolkit.IPUCompiler.DISABLE_PRINT[] = true

IPUToolkit.IPUCompiler.@ipuprintln — Macro

@ipuprint(xs...)
@ipuprintln(xs...)

Print a textual representation of values xs to standard output from the IPU. The functionality builds on @ipuprintf, and is intended as a more use friendly alternative of that API. However, that also means there's only limited support for argument types, handling 16/32/64 signed and unsigned integers, 32 and 64-bit floating point numbers, Cchars and pointers. For more complex output, use @ipuprintf directly.

Limited string interpolation is also possible:

    @ipuprint("Hello, World ", 42, "\n")
    @ipuprint "Hello, World $(42)\n"

Printing can be completely disabled by setting

IPUToolkit.IPUCompiler.DISABLE_PRINT[] = true

IPUToolkit.IPUCompiler.@ipushow — Macro

@ipushow(ex)

IPU analog of Base.@show. It comes with the same type restrictions as @ipuprintf.

@ipushow x

Printing can be completely disabled by setting

IPUToolkit.IPUCompiler.DISABLE_PRINT[] = true

IPUToolkit.IPUCompiler.@ipushowcycles — Macro

@ipushowcycles ex

Print from inside a codelet the expression ex and the number of cycles spent to compute it. This is useful when benchmarking multiple expression, to identify their contributions more easily. The corresponding time can be obtained by dividing the number of cycles by the clock frequency of the the tile, which you can get with Poplar.TargetGetTileClockFrequency(target) outside of the codelet.

See also @ipucycles, @ipuelapsed.

This macro can be made no-op completely by setting

IPUToolkit.IPUCompiler.DISABLE_PRINT[] = true

Options

IPUToolkit.IPUCompiler.DISABLE_PRINT — Constant

IPUToolkit.IPUCompiler.DISABLE_PRINT::Base.RefValue{Bool}

Global constant which controls whether printing through the various @ipuprint* macros should be disabled or not. You may want to completely disable printing for production runs, to avoid the cost of printing on the device, but keep it enabled during development.

Examples:

IPUToolkit.IPUCompiler.DISABLE_PRINT[] = false # Do not disable printing, this is the default.
IPUToolkit.IPUCompiler.DISABLE_PRINT[] = true  # Disable printing, the `@ipuprint*` macros are no-op.

IPUToolkit.IPUCompiler.KEEP_LLVM_FILES — Constant

IPUToolkit.IPUCompiler.KEEP_LLVM_FILES::Base.RefValue{Bool}

Option to control whether to keep in the current directory the files with the LLVM Intermediate Representation (IR) generated for the codelets.

IPUToolkit.IPUCompiler.POPC_FLAGS — Constant

IPUToolkit.IPUCompiler.POPC_FLAGS::Base.RefValue{Cmd}

Options to pass to the popc compiler to compile the code.

IPUToolkit.IPUCompiler.PROGRESS_SPINNER — Constant

IPUToolkit.IPUCompiler.PROGRESS_SPINNER::Base.RefValue{Bool}

Option to control whether to display a spinner to show progress during compilation of IPU codelets.