Writing codelets in Julia

The IPUToolkit.IPUCompiler submodule allows you to write codelets for the IPU in Julia. Codelets are defined with the IPUCompiler.@codelet macro, and then you can use them inside a program, written using the interface to the Poplar SDK described before. This mechanism uses the GPUCompiler.jl package, which is a generic framework for generating LLVM IR code for specialised targets, not limited to GPUs despite the historical name. The IPUCompiler.@codelet macro takes two arguments:

the graph to which to add the codelet with the Poplar.GraphAddCodelets function

the function definition of the kernel you want to compile for the device.

The arguments of a codelet function have to be VertexVector{T,S}, a subtype of AbstractVector{T} which represents a vertex vector, passed as argument to a codelet. The parameters of VertexVector{T,S} are

T: the type of the elements of the vector;

S: the scope of the vector in the codelet, In, Out, or InOut.

Examples of codelets written in Julia are shown in the files examples/main.jl, examples/adam.jl, and examples/pi.jl.

The code inside a codelet has the same limitations as all the compilation models based on GPUCompiler.jl:

the code has to be statically inferred and compiled, dynamic dispatch is not admitted;

you cannot use functionalities which require the Julia runtime, most notably the garbage collector;

you cannot call into any other external binary library at runtime, for example you cannot call into a BLAS library.

After defining a codelet with @codelet you can add a vertex calling this codelet to the graph with the function add_vertex, which also allows controlling the tile mapping in a basic way. Read its docstring for more details.

During compilation of codelets a spinner is displayed to show the progress, as this step can take a few seconds for each codelet to be generated. This can be disabled by setting IPUCompiler.PROGRESS_SPINNER:

IPUCompiler.PROGRESS_SPINNER[] = false

IPUToolkit.IPUCompiler.add_vertex — Function

add_vertex(graph::Poplar.GraphAllocated,
           compute_set_or_program::Union{Poplar.ComputeSetAllocated, Poplar.ProgramSequenceAllocated},
           [tiles::Union{Integer,AbstractVector{<:Integer}},]
           codelet::Function,
           args::Poplar.TensorAllocated...) -> Nothing

Add the codelet function codelet created with @codelet to graph, using the tensors args as arguments. The function codelet must have exactly one method, no more, no less. The second argument can be either the program or the compute set to which to add the new vertex/vertices. If a program is passed, a new compute set will be automatically created.

add_vertex also evenly maps all tensors and vertices across all tiles, which can be either a single tile ID or an AbstractVector of IDs and defaults to single tile 0 if this argument is omitted. Note that all argument tensors args must be longer than or equal to the number of tiles. If you want to have better control over tile mapping, do not use add_vertex.

IPUToolkit.IPUCompiler.PROGRESS_SPINNER — Constant

IPUToolkit.IPUCompiler.PROGRESS_SPINNER::Base.RefValue{Bool}

Option to control whether to display a spinner to show progress during compilation of IPU codelets.

Printing

Inside codelets you can print text and value of variables using the macros @ipuprintf, @ipuprint, @ipuprintln, and @ipushow. These macros are useful for debugging purposes but printing inside a codelet might incur performance penalty. To completely disable all printing and make these macros no-op you can set IPUCompiler.DISABLE_PRINT:

IPUCompiler.DISABLE_PRINT[] = true

IPUToolkit.IPUCompiler.@ipuprintf — Macro

@ipuprintf("%Fmt", args...)

Print a formatted string in device context on the host standard output.

Note that this is not a fully C-compliant printf implementation.

Also beware that it is an untyped, and unforgiving printf implementation. Type widths need to match, eg. printing a 64-bit Julia integer requires the %ld formatting string.

Printing can be completely disabled by setting

IPUToolkit.IPUCompiler.DISABLE_PRINT[] = true

IPUToolkit.IPUCompiler.@ipuprint — Macro

@ipuprint(xs...)
@ipuprintln(xs...)

Print a textual representation of values xs to standard output from the IPU. The functionality builds on @ipuprintf, and is intended as a more use friendly alternative of that API. However, that also means there's only limited support for argument types, handling 16/32/64 signed and unsigned integers, 32 and 64-bit floating point numbers, Cchars and pointers. For more complex output, use @ipuprintf directly.

Limited string interpolation is also possible:

    @ipuprint("Hello, World ", 42, "\n")
    @ipuprint "Hello, World $(42)\n"

Printing can be completely disabled by setting

IPUToolkit.IPUCompiler.DISABLE_PRINT[] = true

IPUToolkit.IPUCompiler.@ipuprintln — Macro

@ipuprint(xs...)
@ipuprintln(xs...)

Print a textual representation of values xs to standard output from the IPU. The functionality builds on @ipuprintf, and is intended as a more use friendly alternative of that API. However, that also means there's only limited support for argument types, handling 16/32/64 signed and unsigned integers, 32 and 64-bit floating point numbers, Cchars and pointers. For more complex output, use @ipuprintf directly.

Limited string interpolation is also possible:

    @ipuprint("Hello, World ", 42, "\n")
    @ipuprint "Hello, World $(42)\n"

Printing can be completely disabled by setting

IPUToolkit.IPUCompiler.DISABLE_PRINT[] = true

IPUToolkit.IPUCompiler.@ipushow — Macro

@ipushow(ex)

IPU analog of Base.@show. It comes with the same type restrictions as @ipuprintf.

@ipushow x

Printing can be completely disabled by setting

IPUToolkit.IPUCompiler.DISABLE_PRINT[] = true

IPUToolkit.IPUCompiler.DISABLE_PRINT — Constant

IPUToolkit.IPUCompiler.DISABLE_PRINT::Base.RefValue{Bool}

Global constant which controls whether printing through the various @ipuprint* macros should be disabled or not. You may want to completely disable printing for production runs, to avoid the cost of printing on the device, but keep it enabled during development.

Examples:

IPUToolkit.IPUCompiler.DISABLE_PRINT[] = false # Do not disable printing, this is the default.
IPUToolkit.IPUCompiler.DISABLE_PRINT[] = true  # Disable printing, the `@ipuprint*` macros are no-op.

Benchmarking

To benchmark expressions inside codelets you can use the macros @ipucycles, @ipushowcycles, and @ipuelapsed, which report the number of cycles spent in the wrapped expression. They are similar to Julia's @time, @showtime, and @elapsed macros, but report the number of cycles, as the clockspeed of tiles cannot be easily obtained inside a codelet. The corresponding time can be obtained by dividing the number of cycles by the clock frequency of the the tile, which you can get with Poplar.TargetGetTileClockFrequency(target) outside of the codelet, and should usually be 1.330 GHz or 1.850 GHz depending on the model of your IPU. The printing macros @ipucycles and @ipushowcycles can be made completely no-op by setting IPUCompiler.DISABLE_PRINT.

Warning

Timing of expressions taking longer than typemax(UInt32) / tile_clock_frequency (about 2 or 3 seconds depending on your IPU model) is unreliable because the difference between the starting and the ending cycle counts would overflow.

IPUToolkit.IPUCompiler.@ipucycles — Macro

@ipucycles ex
@ipucycles "description" ex

Print from inside a codelet the number of cycles spent to compute the expression ex. The corresponding time can be obtained by dividing the number of cycles by the clock frequency of the the tile, which you can get with Poplar.TargetGetTileClockFrequency(target) outside of the codelet. The optional argument description, a literal String, can be used to print also a label to identify the timed expression. A label is added automatically by @ipushowcycles.

See also @ipuelapsed.

This macro can be made no-op completely by setting

IPUToolkit.IPUCompiler.DISABLE_PRINT[] = true

IPUToolkit.IPUCompiler.@ipushowcycles — Macro

@ipushowcycles ex

Print from inside a codelet the expression ex and the number of cycles spent to compute it. This is useful when benchmarking multiple expression, to identify their contributions more easily. The corresponding time can be obtained by dividing the number of cycles by the clock frequency of the the tile, which you can get with Poplar.TargetGetTileClockFrequency(target) outside of the codelet.

See also @ipucycles, @ipuelapsed.

This macro can be made no-op completely by setting

IPUToolkit.IPUCompiler.DISABLE_PRINT[] = true

IPUToolkit.IPUCompiler.@ipuelapsed — Macro

@ipuelapsed ex

Return number of cycles spent to compute the expression ex. The corresponding time can be obtained by dividing the number of cycles by the clock frequency of the the tile, which you can get with Poplar.TargetGetTileClockFrequency(target) outside of the codelet.

See also @ipucycles, @ipushowcycles.

Other options

Other options related to codelet generation are:

IPUToolkit.IPUCompiler.KEEP_LLVM_FILES — Constant

IPUToolkit.IPUCompiler.KEEP_LLVM_FILES::Base.RefValue{Bool}

Option to control whether to keep in the current directory the files with the LLVM Intermediate Representation (IR) generated for the codelets.

IPUToolkit.IPUCompiler.POPC_FLAGS — Constant

IPUToolkit.IPUCompiler.POPC_FLAGS::Base.RefValue{Cmd}

Options to pass to the popc compiler to compile the code.

Domain-Specific Language: @ipuprogram

The IPUCompiler.@ipuprogram macro provides a very simple and limited DSL to automatically generate most of the boilerplate code needed when writing an IPU program. You can do very little with this DSL, which is mainly a showcase of Julia's meta-programming capabilities. A fully commented examples of use of the @ipuprogram macro is available in the examples/dsl.jl file.