Interfacing the Poplar SDK

A quick example of use of the Poplar SDK functionalities, available in the IPUToolkit.Poplar submodule:

julia> using IPUToolkit.Poplar

julia> dm = Poplar.DeviceManager();

julia> Int(Poplar.DeviceManagerGetNumDevices(dm))
129

julia> device = Poplar.get_ipu_device();
[ Info: Trying to attach to device 0...
[ Info: Successfully attached to device 0

julia> Int(Poplar.DeviceGetId(device))
0

julia> Poplar.DeviceDetach(device)

A couple of basic examples of programs running on the IPU written using the interface to the Poplar SDK are available in the files examples/tutorial1.jl and examples/tutorial2.jl.

We automatically generate the bindings of the Poplar SDK using Clang.jl and CxxWrap.jl. There is not automatic documentation at the moment, but functions can be accessed from the Poplar submodule. Also, the IPUToolkit.Poplar submodule wraps a subset of the functionalities available in the Poplar SDK, do not expect to be able to use all functionalities. Remember that Julia does not use class-based object-oriented programming, class instances will usually be first arguments of the methods you want to use.

Function naming convention and signature is usually as follows:

class name in CamelCase, followed by method name also in CamelCase. Note that first letter of method name is always uppercase in this naming convention, even if it is lowercase in the Poplar SDK. For example, the method getNumDevices of the DeviceManager class can be accessed in the Poplar submodule with Poplar.DeviceManagerGetNumDevices;

the first argument of the function is the class instance. For example, to use the Julia function Poplar.DeviceManagerGetNumDevices, you need to pass as first argument an instance of DeviceManager;

the following arguments are the same as in the method you want to use in the SDK. For example, the method getNumDevices of the DeviceManager class doesn't take any argument, so the Julia function Poplar.DeviceManagerGetNumDevices will take an instance of DeviceManager as only argument.

Convenient methods

In addition to this, for some functions (e.g. EngineWriteTensor, EngineConnectStream, EngineReadTensor) we provide more user-friendly methods where the last argument can be a Julia's Array, without having to pass additional arguments for pointers or array size. Furthermore, the custom functions Poplar.get_ipu_device and Poplar.get_ipu_devices can be used to access one more IPU devices, as shown in the example above.

Another function for which we provide a convenient method is Poplar.GraphAddConstant:

Poplar.GraphAddConstant(graph, host_array)

adds the host_array (a plain standard Julia Array living on the host) to graph, automatically inferring from host_array the type and the shape of the tensor in the graph.

You can slice a tensor with the usual Julia notation tensor[index1:index2], this corresponds to a call to Tensor.slice(index1, index2+1).

similar can be used to add to graph a tensor with the same shape and optionally element type as tensor.

Base.similar — Method

similar(
    graph::Poplar.Graph,
    tensor::Poplar.TensorAllocated,
    [type::DataType],
    [debug::String]
) -> Poplar.TensorAllocated

Adds to graph a variable tensor with the same shape as tensor. If a type (this is a Julia type, like Float32 or Int32) argument is not passed, the same element type as tensor will be automatically used. An optional debug context can also be passed, as a String.

IPUToolkit.Poplar.get_ipu_device — Function

get_ipu_device(hint::Union{AbstractVector{<:Integer},Integer}=0)

Similar to get_ipu_devices, but request exactly one IPU device. If it can attach to a device, return that pointer only (not in a vector, like get_ipu_devices), otherwise return nothing. You can release the device with Poplar.DeviceDetach(device).

The optional argument hint suggests to which device IDs to try and attach. It can have different types:

if of type Integer, try to attach to one device, starting from the one with index hint. The default is hint=0;

if of type AbstractVector, try to attach to a device from that list of IDs.

IPUToolkit.Poplar.get_ipu_devices — Function

get_ipu_devices(n::Int, hint::Union{AbstractVector{<:Integer},Integer}=0)

Try to attach to n IPU devices, returns a vector of the pointers to the devices successfully attached to. You can release them with Poplar.DeviceDetach (note that this function takes a single pointer as input, so you have to use broadcasting Poplar.DeviceDetach.(devices) to release a vector of pointers).

The second optional argument hint suggests to which device IDs to try and attach. It can have different types:

if of type Integer, try to attach to n devices, starting from the one with index hint. The default is hint=0;

if of type AbstractVector, try to attach to n devices from that list of IDs.

See get_ipu_device for requesting exactly one IPU device.